- Use gogo for marshalling with better Go code generation 	- To properly cast big.Int, we use a custom patch and detail how to use it 	- We implement the following using protobufs: 		- curve.Point, curve.Scalar, paillier.Ciphertext, polynomial.Exponent, pedersen.Parameters, and all zero-knownledge proofs - hash.Hash only reports errors when necessary, and cleanup unused functions - Perform better validation overall with integers mod N - Implement curve.Scalar and curve.Point using the base types from the decred module 	- All marshalling now uses the compressed encoding of Point as to perform validation 	- Identity points are no longer valid - polynomial.Exponent now has a flag indicating whether the constant coefficient is the identity, which prevents the need of marshalling identity points - session.Session is now an interface and is implemented as 3 structs for Keygen, Refresh and Sign - ZK 	- use protobuf everywhere 	- much better validation of proofs 	- added comments 	- Schnorr now supports multiple points simultaneously - protocol.Sign: 	- better MtA design
- protocols:
	- WIP: rename all fields to use capital letters to allow Marshalling of state for error reporting in case of failure

+++ pkg/party/idslice_test.go
@@ -19,5 +19,5 @@ func TestIDSlice_Lagrange(t *testing.T) {
 	for _, c := range coefs {
 		sum.Add(sum, c)
 	}
-	assert.True(t, sum.Equal(curve.NewScalar().SetInt64(1)))
+	assert.True(t, sum.Equal(curve.NewScalar().SetUInt32(1)))
 }
